+++ pyviews/__init__.py
@@ -1,3 +1,3 @@
 """Package for creating tkinter applications in declarative way."""
 
-__version__ = '3.0.1'
+__version__ = '3.1.0'