+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 
-## Dev
+## 3.1.0
 
 - added containers(Container, View, If, For)
 - added presenter